feat: add new tab for map in main screen
