+++ src/components/map/SelectionPin.tsx
